move app_to_dict[...] to application_commands.py
